+++ src/test/java/io/github/astrapi69/swing/tree/panel/NodeModelBean.java
@@ -42,7 +42,7 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 @Builder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class AddNodeModelBean
+public class NodeModelBean
 {
 	boolean node;
 	String name;