+++ CHANGELOG.md
@@ -9,6 +9,8 @@ ADDED:
 - new class GenericTreeElement created
 - new class GenericTreeNodeCellRenderer created
 - new class TreeNodeGenericTreeElementWithContentPanel created
+- new class TreeNodeGenericTreeElementWithGenericContentPanel created
+- new method for get the selected user object from a given JTree object
 - new test dependency for icons org.freedesktop.tango:tango-icon-theme in patch version 0.8.90 
 - new test dependency for icons io.github.astrapi69:silk-icon-theme in major version 1